+++ src/main.rs
@@ -1,7 +1,9 @@
 use pre_commit_sort::{PreCommit, PreCommitConfig, PreCommitHooks};
+use std::env;
 
 fn main() -> anyhow::Result<()> {
-    PreCommitConfig::main()?;
-    PreCommitHooks::main()?;
+    let install = env::args().nth(1) == Some("-i".to_string());
+    PreCommitConfig::main(install)?;
+    PreCommitHooks::main(false)?;
     Ok(())
 }