+++ src/pre_commit_hooks.rs
@@ -5,7 +5,7 @@ pub type PreCommitHooks = Vec<DeclareHook>;
 impl PreCommit for PreCommitHooks {
     const PATH: &'static str = ".pre-commit-hooks.yaml";
 
-    fn process(&mut self) {
+    fn process(&mut self, _install: bool) {
         self.sort();
     }
 }